requirements.txt
ansicolors==1.1.8
APScheduler==3.6.3
bidict==0.18.3
biosppy==0.6.1
bitstring==3.1.6
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==3.7.4
cycler==0.10.0
docopt==0.6.2
et-xmlfile==1.0.1
Flask==1.0.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.8
future==0.17.1
h5py==2.10.0
html5lib==1.0.1
idna==2.8
img2pdf==0.3.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mongoengine==0.18.2
nose==1.3.7
numpy==1.16.1
opencv-python==4.1.1.26
openpyxl==3.0.2
pandas==0.25.1
Pillow==5.4.1
pydicom==1.2.2
pymodm==0.4.1
pymongo==3.7.2
pyparsing==2.3.1
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
reportlab==3.5.13
requests==2.21.0
scikit-learn==0.21.3
scipy==1.2.1
shortuuid==0.5.0
six==1.12.0
sklearn==0.0
tzlocal==2.0.0
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
wfdb==2.2.1
WSGIserver==1.3
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
